backend/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
neo4j
pgvector
sqlalchemy
langgraph
langfuse
python-dotenv
pydantic
requests
python-docx
PyPDF2
pytest
